Fixed a bug in MCOIndexSet::adRange()
If the indexes were added in an order that required the ranges before and after the insertion point to be merged then the second merge would fail. For instance, adding  10 then 11 then 12 would result in an index set defined as 10 - 12 (correct), but adding 12 then 10 then 11 would result in an index set defined as 10-11, 12. These would behave the same way, but do not compare as equal.